// src/app/models/user-api.interface.ts
import { Address } from './address.interface';
import { Company } from './company.interface';

export interface UserApiInterface {
  id: string;
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}